% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl_aewma.R
\name{arl_aewma}
\alias{arl_aewma}
\title{ARL value of the Adaptive EWMA control chart (with p-values)}
\usage{
arl_aewma(alpha, w = 50, nsimul, shift = 0)
}
\arguments{
\item{alpha}{a real number; the level of significance}

\item{w}{an integer; the sample size needed to reach steady-state, default value is 50.}

\item{nsimul}{an integer; the number of replications}

\item{shift}{a real number; shift size. If shift=0, IC ARL is returned, else OC ARL is returned. Default value is 0.0.}
}
\value{
ARL
}
\description{
ARL value of the Adaptive EWMA control chart (with p-values)
}
\details{
Estimates the ARL for the AEWMA chart (with p-values)
}
\examples{
arl_aewma(0.025,50,100,0.0)
}
\author{
Samuel Anyaso-Samuel
}
