% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl_aewma_max.R
\name{arl_aewma_max}
\alias{arl_aewma_max}
\title{ARL value of the one-sided(max) Adaptive EWMA control chart (without p-values)}
\usage{
arl_aewma_max(alpha, w = 50, nsimul, shift = 0)
}
\arguments{
\item{alpha}{a real number; the level of significance}

\item{w}{an integer; the sample size needed to reach steady-state, default value is 50.}

\item{nsimul}{an integer; the number of replications}

\item{shift}{a real number; shift size. If shift=0, IC ARL is returned, else OC ARL is returned. Default value is 0.0.}
}
\value{
ARL, count of success iterations
}
\description{
ARL value of the one-sided(max) Adaptive EWMA control chart (without p-values)
}
\details{
Estimates the ARL for one-sided(max) the AEWMA chart (without p-values)
}
\examples{
arl_aewma_max(0.025,50,100,0.0)
}
\author{
Samuel Anyaso-Samuel
}
