% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats_arl_aewma.R
\name{ats_arl_aewma}
\alias{ats_arl_aewma}
\title{ARL & ATS of the Adaptive EWMA control chart.}
\usage{
ats_arl_aewma(alpha, w = 50, nsimul, a, lambda, b, shift = 0)
}
\arguments{
\item{alpha}{a real number; the level of significance}

\item{w}{an integer; the sample size needed to reach steady-state, default value is 50.}

\item{nsimul}{an integer; the number of replications}

\item{a}{a real number; a parameter in the dynamic sampling interval function}

\item{lambda}{a real number; lambda parameter in the dynamic sampling interval function}

\item{b}{a real number; b parameter in the dynamic sampling interval function}

\item{shift}{a real number;  shift size. If shift=0, IC ARL/ATS is returned, else OC ARL/ATS is returned. Default value is 0.0.}
}
\value{
ATS and ARL[when d(.) = 1]
}
\description{
ARL & ATS of the Adaptive EWMA control chart.
}
\details{
Estimates the ARL & ATS for the AEWMA chart using the bootstrap approach.
}
\examples{
ats_arl_aewma(0.025,50,100,0,2.5,3.1,0.05)
}
\author{
Samuel Anyaso-Samuel
}
