% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats_arl_aewma_max.R
\name{ats_arl_aewma_max}
\alias{ats_arl_aewma_max}
\title{ARL & ATS of the one-sided(max) Adaptive EWMA control chart.}
\usage{
ats_arl_aewma_max(alpha, w = 50, nsimul, repl = 1e+06, a, lambda, b, shift = 0)
}
\arguments{
\item{alpha}{a real number; the level of significance}

\item{w}{an integer; the sample size needed to reach steady-state, default value is 50.}

\item{nsimul}{an integer; the number of simulations}

\item{repl}{an integer; the number of replications for estimating the emprical distribution}

\item{a}{a real number; a parameter in the dynamic sampling interval function}

\item{lambda}{a real number; lambda parameter in the dynamic sampling interval function}

\item{b}{a real number; b parameter in the dynamic sampling interval function}

\item{shift}{a real number;  shift size. If shift=0, IC ARL/ATS is returned, else OC ARL/ATS is returned. Default value is 0.0.}
}
\value{
ATS and ARL[when d(.) = 1] and count of successful iterations
}
\description{
ARL & ATS of the one-sided(max) Adaptive EWMA control chart.
}
\details{
Estimates the ARL & ATS for the one-sided(max) AEWMA chart using the bootstrap approach.
}
\examples{
ats_arl_aewma_max(0.025,50,100,1000,0,2.5,3.1,0.05)
}
\author{
Samuel Anyaso-Samuel
}
