% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empr_aewma.R
\name{empr_aewma}
\alias{empr_aewma}
\title{Empirical Distribution of the Adaptive EMWA control chart}
\usage{
empr_aewma(x, nsimul, w)
}
\arguments{
\item{x}{a numeric vector; observed IC data values}

\item{nsimul}{an integer; number of bootstrap simulations}

\item{w}{an integer; the sample size needed to reach steady-state}
}
\value{
a numeric vector of the empirical distribution of the proposed AEWMA chart
}
\description{
Empirical Distribution of the Adaptive EMWA control chart
}
\details{
Estimates the in-control empirical distribution of the adaptive EWMA chart.
}
\examples{
empr_aewma(rnorm(100),1000, 50)
}
\author{
Samuel Anyaso-Samuel
}
