% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empr_aewma_max.R
\name{empr_aewma_max}
\alias{empr_aewma_max}
\title{Empirical Distribution of the one-sided(max) Adaptive EMWA control chart}
\usage{
empr_aewma_max(nsimul, w, x)
}
\arguments{
\item{nsimul}{an integer; number of bootstrap simulations}

\item{w}{an integer; the sample size needed to reach steady-state}

\item{x}{a numeric vector; observed IC data values}
}
\value{
a numeric vector of the empirical distribution of the proposed AEWMA chart
}
\description{
Empirical Distribution of the one-sided(max) Adaptive EMWA control chart
}
\details{
Estimates the in-control empirical distribution of the one-sided(max) adaptive EWMA chart.
}
\examples{
empr_aewma_max(1000, 50, rnorm(100))
}
\author{
Samuel Anyaso-Samuel
}
