% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pVal.R
\name{pVal}
\alias{pVal}
\title{Computes the p-value from the empirical distribution of the statistic}
\usage{
pVal(empr_dist, obsStat)
}
\arguments{
\item{empr_dist}{a numeric vector; the empirical distribution}

\item{obsStat}{a real number; the observed test statistic}
}
\value{
p-value (right-tailed) of the observed test statistic
}
\description{
Computes the p-value from the empirical distribution of the statistic
}
\details{
Estimates the p-value (Right-tailed) for the AEWMA chart proposed by Haq et al
}
\examples{
pVal(c(0.25,-0.36,-0.54,0.89), 0.64)
}
\author{
Samuel Anyaso-Samuel
}
